app: Move more store related functions to ee
