+++ tracecat/dsl/action.py
@@ -11,7 +11,6 @@ from temporalio.exceptions import ApplicationError
 from tracecat.contexts import ctx_logger, ctx_run
 from tracecat.dsl.common import context_locator
 from tracecat.dsl.models import ActionStatement, DSLTaskErrorInfo, RunActionInput
-from tracecat.ee.store.models import StoreObjectPtr
 from tracecat.executor.client import ExecutorClient
 from tracecat.logger import logger
 from tracecat.registry.actions.models import RegistryActionValidateResponse
@@ -19,7 +18,7 @@ from tracecat.types.auth import Role
 from tracecat.types.exceptions import RegistryActionError
 
 
-def _contextualize_message(
+def contextualize_message(
     task: ActionStatement,
     msg: str | BaseException,
     *,
@@ -111,7 +110,7 @@ class DSLActivities:
             # We only expect RegistryActionError to be raised from the registry client
             kind = e.__class__.__name__
             msg = str(e)
-            err_locator = _contextualize_message(task, msg, attempt=attempt)
+            err_locator = contextualize_message(task, msg, attempt=attempt)
             act_logger.error(
                 "Application exception occurred", error=msg, detail=e.detail
             )
@@ -132,7 +131,7 @@ class DSLActivities:
                 attempt=attempt,
             )
             raise ApplicationError(
-                _contextualize_message(task, e.message, attempt=attempt),
+                contextualize_message(task, e.message, attempt=attempt),
                 err_info,
                 non_retryable=e.non_retryable,
                 type=e.type,
@@ -150,60 +149,8 @@ class DSLActivities:
                 attempt=attempt,
             )
             raise ApplicationError(
-                _contextualize_message(task, raw_msg, attempt=attempt),
+                contextualize_message(task, raw_msg, attempt=attempt),
                 err_info,
                 type=kind,
                 non_retryable=True,
-            ) from e
-
-    @staticmethod
-    @activity.defn
-    async def run_action_with_store_activity(
-        input: RunActionInput, role: Role
-    ) -> StoreObjectPtr:
-        """
-        Run an action in store mode.
-
-        Note
-        ----
-        We must only pass store object pointers across activities/workflows.
-        Do not pass the objects themselves as they are too heavy to be transferred over the boundary.
-        DSLContext should just keep track of the references to the objects, not the objects themselves.
-        """
-        ctx_run.set(input.run_context)
-        log = logger.bind(
-            task_ref=input.task.ref,
-            action_name=input.task.action,
-            wf_id=input.run_context.wf_id,
-            role=role,
-            environment=input.run_context.environment,
-        )
-        ctx_logger.set(log)
-
-        attempt = activity.info().attempt
-        log.info(
-            "Run action activity",
-            task=input.task,
-            attempt=attempt,
-            retry_policy=input.task.retry_policy,
-        )
-
-        # Add a delay
-        if input.task.start_delay > 0:
-            log.info("Starting action with delay", delay=input.task.start_delay)
-            await asyncio.sleep(input.task.start_delay)
-
-        try:
-            # Delegate to the registry client
-            client = ExecutorClient(role=role)
-            handle = await client.run_action_store_backend(input)
-            return handle.to_pointer()
-        except Exception as e:
-            # Now that we return ActionRefHandle, these are transient errors
-            kind = e.__class__.__name__
-            raise ApplicationError(
-                _contextualize_message(input.task, e, attempt=attempt),
-                DSLTaskErrorInfo(
-                    ref=input.task.ref, message=str(e), type=kind, attempt=attempt
-                ),
             ) from e
